+++ src/__tests__/vault.test.ts
@@ -10,6 +10,7 @@ jest.setTimeout(3000000);
 
 describe("Testing vault functions", () => {
   let vaultId: string;
+  let publicVaultId: string;
 
   beforeAll(async () => {
     akord = await initInstance();
@@ -86,5 +87,34 @@ describe("Testing vault functions", () => {
 
     const vault = await akord.vault.get(vaultId);
     expect(vault.status).toEqual(status.ACTIVE);
+  });
+
+  it("should create vault with tags", async () => {
+    const name = faker.random.words();
+
+    const tag1 = faker.random.word();
+    const tag2 = faker.random.word();
+    const vault = await akord.vault.create(name, { public: true, tags: [tag1, tag2] });
+
+    expect(vault.tags).toBeTruthy();
+    expect(vault.tags?.length).toEqual(2);
+    expect(vault.tags).toContain(tag1);
+    expect(vault.tags).toContain(tag2);
+
+    publicVaultId = vault.id;
+  });
+
+  it("should update vault metadata", async () => {
+    const description = faker.random.words();
+
+    const tag1 = faker.random.word();
+    const tag2 = faker.random.word();
+    const vault = await akord.vault.update(publicVaultId, { description: description, tags: [tag1, tag2] });
+
+    expect(vault.description).toEqual(description);
+    expect(vault.tags).toBeTruthy();
+    expect(vault.tags?.length).toEqual(2);
+    expect(vault.tags).toContain(tag1);
+    expect(vault.tags).toContain(tag2);
   });
 });