+++ src/api/akord-api.ts
@@ -125,6 +125,7 @@ export default class AkordApi extends Api {
       .public(tx.public)
       .name(tx.name)
       .description(tx.description)
+      .tags(tx.tags)
       .keys(tx.keys)
       .autoExecute(this.autoExecute)
       .createVault();
@@ -137,6 +138,7 @@ export default class AkordApi extends Api {
       .resourceId(tx.id)
       .name(tx.name)
       .description(tx.description)
+      .tags(tx.tags)
       .status(tx.status)
       .autoExecute(this.autoExecute)
       .updateVault();
@@ -176,6 +178,7 @@ export default class AkordApi extends Api {
       .keys(tx.keys)
       .name(tx.name)
       .encPrivateKey(tx.encPrivateKey)
+      .ownerAccess(tx.ownerAccess)
       .allowedStorage(tx.allowedStorage)
       .allowedPaths(tx.allowedPaths)
       .autoExecute(this.autoExecute)