+++ src/types/transaction.ts
@@ -16,13 +16,15 @@ export interface CreateVaultTxPayload {
   name?: string,
   description?: string
   public?: boolean,
-  keys?: Array<EncryptedVaultKeyPair>
+  keys?: Array<EncryptedVaultKeyPair>,
+  tags?: string[]
 }
 
 export interface UpdateVaultTxPayload {
   id: string,
   name?: string,
   description?: string
+  tags?: string[]
   status?: string
 }
 
@@ -61,6 +63,7 @@ export interface CreateMembershipTxPayload {
   expiresAt?: number,
   keys?: EncryptedVaultKeyPair[],
   encPrivateKey?: string,
+  ownerAccess?: string,
   allowedStorage?: number,
   allowedPaths?: AllowedPaths
 }